Ton of bug fix

+++ MashGamemodeLibrary/Audio/Containers/ISyncedAudioContainer.cs
@@ -4,7 +4,7 @@ namespace MashGamemodeLibrary.Audio.Containers;
 
 public interface ISyncedAudioContainer : IAudioContainer
 {
-    List<int> AudioHashes { get; }
-    int? GetAudioHash(string name);
-    void RequestClip(int hash, Action<AudioClip?> onClipReady);
+    List<ulong> AudioHashes { get; }
+    ulong? GetAudioHash(string name);
+    void RequestClip(ulong hash, Action<AudioClip?> onClipReady);
 }
